+++ cli/diff/data_naive.go
@@ -2,10 +2,13 @@ package diff
 
 import (
 	"context"
-	"github.com/samber/lo"
-	"golang.org/x/sync/errgroup"
+	"fmt"
 	"io"
 	"slices"
+
+	"github.com/neilotoole/sq/libsq/core/lg/lgm"
+	"github.com/samber/lo"
+	"golang.org/x/sync/errgroup"
 
 	"github.com/neilotoole/sq/cli/run"
 	"github.com/neilotoole/sq/libsq/core/ioz/contextio"
@@ -34,7 +37,7 @@ func execSourceDataDiff(ctx context.Context, ru *run.Run, cfg *Config, sd1, sd2 
 	docs := make([]*HunkDoc, len(allTblNames))
 	defer func() {
 		for i := range docs {
-			lg.WarnIfCloseError(log, "Close diff doc", docs[i])
+			lg.WarnIfCloseError(log, lgm.CloseDiffDoc, docs[i])
 		}
 	}()
 
@@ -49,7 +52,7 @@ func execSourceDataDiff(ctx context.Context, ru *run.Run, cfg *Config, sd1, sd2 
 		td2 := &tableData{src: sd2.src, tblName: tblName}
 		td2.tblMeta = sd2.srcMeta.Table(tblName)
 
-		cmdTitle := []byte(cfg.Colors.CmdTitle.Sprintf("sq diff --data %s %s", td1.String(), td2.String()))
+		cmdTitle := Titlef(cfg.Colors, fmt.Sprintf("sq diff --data %s %s", td1.String(), td2.String()))
 		doc := NewHunkDoc(cmdTitle, NewDocHeader(cfg.Colors, td1.String(), td2.String()))
 		docs[i] = doc
 		execFns[i] = func() error {
